enhancement: can only delete a user with admin privledges or when signed in as that user

+++ src/main/java/com/ryansstore/store/users/UserService.java
@@ -6,6 +6,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.data.domain.Sort;
+import com.ryansstore.store.authentication.AuthService;
 import lombok.AllArgsConstructor;
 import java.util.Set;
 import java.util.List;
@@ -17,6 +18,7 @@ public class UserService {
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
     private final AuthenticationManager authManager;
+    private final AuthService authService;
 
     public UserDto register(UserRegisterRequest request) {
         if(userRepository.existsByEmail(request.getEmail()))
@@ -59,6 +61,10 @@ public class UserService {
     }
 
     public ResponseEntity<Void> deleteUser(Long id) {
+        User currentUser = authService.getCurrentUser();
+        if(!currentUser.getId().equals(id) && !currentUser.getRole().equals(Role.ADMIN))
+            throw new UserUnauthorizedDeleteException();
+
         User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
 
         userRepository.delete(user);
